# standdienst-api/static/openapi.yaml
openapi: 3.1.0
info:
  title: Standdienst API
  version: "3.8.0"
  description: |
    REST-API für die Standdienst-Plattform zur Verwaltung von Freiwilligendiensten und Essensspenden.

    **Authentifizierung:** JWT via HttpOnly-Cookie (`access_token`) oder `Authorization: Bearer <token>` Header.

    **CSRF:** Für alle schreibenden Anfragen muss der `X-CSRF-TOKEN` Header mit dem CSRF-Token aus dem `csrf_access_token` Cookie mitgesendet werden.

    **Rollen:**
    - `admin` – Globaler Administrator
    - `organizer` – Instanz-Verantwortlicher
    - `volunteer` – Helfer (instanzgebunden)
servers:
  - url: /api
    description: API-Basis
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /public/{slug}/info:
    get:
      tags: [public]
      summary: Instanz-Info + Branding
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Instanz-Info
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      name: { type: string }
                      slug: { type: string }
                      site_title: { type: string }
                      primary_color: { type: string }
                      logo_url: { type: string, nullable: true }
                      mail_enabled: { type: boolean }
                      has_privacy_policy: { type: boolean }
        "404": { description: Instanz nicht gefunden }
  /public/{slug}/captcha:
    get:
      tags: [public]
      summary: CAPTCHA generieren
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Mathe-CAPTCHA
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      question: { type: string, example: "3 + 5 = ?" }
  /public/{slug}/register:
    post:
      tags: [public]
      summary: Volunteer registrieren
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, captcha_answer]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                captcha_answer: { type: integer }
                consent: { type: boolean }
      responses:
        "201": { description: Registrierung erfolgreich }
        "400": { description: Validierungsfehler oder CAPTCHA falsch }
        "409": { description: E-Mail bereits registriert }
  /public/{slug}/datenschutz:
    get:
      tags: [public]
      summary: Datenschutzerklärung
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: HTML der Datenschutzerklärung
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      html: { type: string }

  # ── Volunteer ──────────────────────────────────────────────────────────
...
  /admin/{slug}/settings:
    get:
      tags: [admin]
      summary: Instanz-Einstellungen abrufen
      security:
        - cookieAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Instanz-Einstellungen }
    put:
      tags: [admin]
      summary: Instanz-Einstellungen aktualisieren
      security:
        - cookieAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Einstellungen gespeichert }
  /admin/settings/global:
    get:
      tags: [admin]
      summary: Globale Einstellungen abrufen
      security:
        - cookieAuth: []
      responses:
        "200": { description: Globale Einstellungen }
    put:
      tags: [admin]
      summary: Globale Einstellungen aktualisieren
      security:
        - cookieAuth: []
      responses:
        "200": { description: Einstellungen gespeichert }
  /admin/settings/mail:
    get:
      tags: [admin]
      summary: Mail-Einstellungen abrufen
      security:
        - cookieAuth: []
      responses:
        "200": { description: SMTP-Konfiguration }
    put:
      tags: [admin]
      summary: Mail-Einstellungen aktualisieren
      security:
        - cookieAuth: []
      responses:
        "200": { description: Mail-Einstellungen gespeichert }
  /admin/settings/mail/test:
    post:
      tags: [admin]
      summary: Test-E-Mail senden
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
      responses:
        "200": { description: Testmail gesendet }
        "503": { description: E-Mail nicht konfiguriert }

  # ── Admin: Backup & Update ─────────────────────────────────────────────
...